Moves manipulators out of controllers folder
