+++ src/controllers/index.ts
@@ -1,5 +1,3 @@
-export { ItemManipulator } from "./itemManipulator";
-export { RoomManipulator } from "./roomManipulator";
-export { AgentManipulator } from "./agentManipulator";
 export { InventoryController } from "./inventoryController";
 export { MovementController } from "./movementController";
+export { SpawnController } from "./spawnController";
