+++ src/controllers/spawnController.ts
@@ -1,8 +1,8 @@
 import { BaseController } from "./baseController";
 import { Agent, Item, Room } from "../models";
-import { ItemManipulator } from "./itemManipulator";
-import { RoomManipulator } from "./roomManipulator";
-import { AgentManipulator } from "./agentManipulator";
+import { ItemManipulator } from "../manipulators/itemManipulator";
+import { RoomManipulator } from "../manipulators/roomManipulator";
+import { AgentManipulator } from "../manipulators/agentManipulator";
 
 export class SpawnController extends BaseController {
