+++ src/manipulators/agentManipulator.ts
