+++ src/manipulators/conversationManipulator.ts
