+++ src/manipulators/index.ts
@@ -1,0 +1,3 @@
+export { ItemManipulator } from "./itemManipulator";
+export { RoomManipulator } from "./roomManipulator";
+export { AgentManipulator } from "./agentManipulator";
