+++ src/manipulators/roomManipulator.ts
